backend/pyproject.toml
[project]
name = "finshield-api"
version = "0.1.0"
description = "FinShield - The Financial Flight Recorder API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.16",
    "sqlalchemy>=2.0.27",
    "python-dateutil>=2.9.0",
    "rapidfuzz>=3.6.1",
    "pillow>=10.0.0",
    "numpy>=1.26.4",
    "opencv-python>=4.10.0.84",
    "pytesseract>=0.3.10",
    "openpyxl>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
